Migrate the habitableZone class to ts
Also, it's no longer a class, but a simple method. The logic for calculating the bounds of the habitable zone has been moved into a "StellarService" which will be the repository for all methods for calculating stellar quantities.
